Add basic (and stupid) kernel heap allocator
This commit is a first stab at allocating memory in the kernel!
As it is right now it is completely broken most likely and the allocator
is trivial and stupid. This is just something to get us started :)

Adds the alloc crate as a dependency, this means we do not have to
implement datastructures like Vec and helpers like Box ourselves. We
just have to register a global allocator.