Kernel: Implement the TSS struct for Interrupt Stacks
Adds a Task Segment Selector struct. This is a legacy struct previously
used for hardware task switching. In 64-bit mode this is no longer
supported and is instead handled in software from the OS.

We, however, still need the TSS to set up separate interrupt stacks.
This enables us to catch for example a kernel stack overflow in the
double fault handler. Otherwise, the double fault handler itself would
trigger a stack overflow leading to a triple fault and a system reset.