Kernel: Implement separate interrupt stacks for double faults
If the kernel has a stack overflow it will trigger a page fault. When
the CPU calls our handler it will first push a InterruptStackFrame on
the stack which again causes a page fault leading to a double fault.
Once again in the double fault handler the same thing happens leading to
a triple fault and a system reset.

To avoid this you can allocate a separate memory buffer that the
interrupt handler can use as it's stack. This requires some outdated
legacy structures in the CPU (a GDT and TSS). This commit enables this
for the double fault handler. This means we can "handle" a kernel stack
overflow.