Test: Add test for kernel stack overflow handling
This integration test tests if we can handle a kernel stack overflow.
You cannot recover from it but you should be able to successfully run
the double fault handler even in that situation.

This is enabled through separate interrupt stacks via the interrupt
stack table and task selector segement in the gdt.

Our first integration test!