mem: add VirtAddr and PhysAddr structs
This is the first start of memory management in the kernel! A very
important aspect of this is differentiating between talking about
physical and virtual addresses. This comes up especially when doing
paging.

Let's use Rust's awesome type system to encode when we are
talking about a physical or virtual address, without any runtime
performance at all!