mem: Add FrameAllocator struct to keep track of used frames
One of the most important jobs of the kernel is keeping tracking of
memory. We need to keep track of which frames (physical pages) are
currently in use. This is the job of the global FrameAllocator.

Currently it's just a stupid bump allocator but it is just something to
get started. Uses the bootinfo supplied by the bootloader to know which
pages are free to use from the start.