meta: remove cargo-xbuild and use built-in alternative!
Relatively recently cargo introduced the `-Zbuild-std` flag which lets
us easily build core/alloc etc from cargo. Previously we used the
cargo-xbuild tool for this. Now we can simply run `cargo run/check/test`
and so on!

A known issue is memcpy, memset, etc not being included in the
`compiler_builtins` build, so for now we need to depend on the rlibc
crate for those implementations..